Added Process Flow Diagram
Added Draw.io Process Flow Diagram and cleaned up some stuff in preparation for larger changes, soon to change Lock Out authentication method before then packaging the CLI module and beginning work on the GUI component
